+++ huffman_coding.cpp
@@ -1,50 +1,18 @@
-#include "huffman.h"
-
-#include <iostream>
-#include <queue>
-#include <stack>
+#include "huffman_coding.h"
+#include "huffman_tree.h"
+#include <string>
 #include <fstream>
+#include <iostream>
 
-const bool _HUFFMAN_DEBUG = 0;
+using namespace Huffman;
 
-/* -- Funciones y estructuras auxiliares -- */
-
-struct BitUdec {
-    unsigned char val : 1;
-};
-
-/* Combina dos nodos como hijos de un nodo padre, */
-/* el cual tiene como frecuencia la suma de la frecuencia de sus hijos */
-Node * combineNodes(Node *, Node *);
-
-/* Clase comparativa: Criterio de comparación, frecuencias */
-struct nodeComparator {
-    bool operator()(const Node * node1, const Node * node2) const {
-        return node1->frequency > node2->frequency;
-    }
-};
-
-/* Obtiene las frecuencias de cada caracter en un texto */
-std::unordered_map<char, int> readFrequencies(std::fstream& textfin);
-
-/* Crea nodos hoja del arbol de Huffman a partir del mapa de frecuencias */
-std::priority_queue<Node *, std::vector<Node*>, nodeComparator> createLeaves(const std::unordered_map<char, int>&);
-
-/* Crea el arbol de Huffman a partir de un texto y su mapa de frecuencias */
-Node * createHuffmanTree(std::unordered_map<char, int> freq);
-
-/* Busca un valor en un arbol binario de Huffman, creando su camino. Recursiva */
-bool searchValue(Node *, char, std::string&);
-
-/* Crea diccionarios de codificación y decodificación a partir de un arbol de Huffman */
-std::pair<std::unordered_map<char, std::string>, std::unordered_map<std::string, char>> getMaps(Node * huff, std::unordered_map<char, int> map);
+const bool _HUFFMAN_CODING_DEBUG = 1;
+size_t Huffman::LastSavedSize = 0;
 
 /* debugging :33 :3 */
 void printbits(unsigned char v);
 
-/* -- Implementación header -- */
-
-std::string HuffmanCoding::str_encode(std::fstream& ifs) 
+std::string Huffman::str_encode(std::fstream& ifs) 
 {
     auto freq_map = readFrequencies(ifs);           
     auto huffman_tree = createHuffmanTree(freq_map);
@@ -64,7 +32,7 @@ std::string HuffmanCoding::str_encode(std::fstream& ifs)
     }
 
     /* DEBUGGING PRINT */
-    if (_HUFFMAN_DEBUG) {
+    if (_HUFFMAN_CODING_DEBUG) {
         for (auto it = coding.begin(); it != coding.end(); it++) {
             auto key = it->first;
             std::cout << "'" << key << "' :     " << coding[key] << std::endl;
@@ -75,7 +43,7 @@ std::string HuffmanCoding::str_encode(std::fstream& ifs)
     return encoded_msg;
 }
 
-std::pair<unsigned int, unsigned char*> HuffmanCoding::bit_encode(std::fstream& ifs) 
+std::pair<unsigned int,unsigned char*> Huffman::bit_encode(std::fstream& ifs) 
 {
     auto freq_map = readFrequencies(ifs);           
     auto huffman_tree = createHuffmanTree(freq_map);
@@ -114,7 +82,7 @@ std::pair<unsigned int, unsigned char*> HuffmanCoding::bit_encode(std::fstream& 
     while (ifs >> std::noskipws >> current_ch) {
         std::string encoded_char = (coding[current_ch]);
 
-        for (int i=0; i < encoded_char.length(); i++) {
+        for (int i=0; i < (int)encoded_char.length(); i++) {
             /* cuando un byte del arreglo se llena, incrementar el indice de encoded_msg para asi pasar al siguiente byte */
             if (bitflip_pos < 0) {
                 EncMsg_idx++;
@@ -129,21 +97,35 @@ std::pair<unsigned int, unsigned char*> HuffmanCoding::bit_encode(std::fstream& 
         }
     }
 
-    /* Debugging print */
-    if (_HUFFMAN_DEBUG) {
+    /* Debugging prints */
+    if (_HUFFMAN_CODING_DEBUG) {
+        std::cout << "Encoding generated: " << std::endl;
+        for (auto it = coding.begin(); it != coding.end(); it++) {
+            auto key = it->first;
+            if (key == '\n')
+                std::cout << "'" << "[NEWLINE]" << "' :     " << coding[key] << std::endl;
+            else if (key == ' ')
+                std::cout << "'" << "[WHITESPACE]" << "' :     " << coding[key] << std::endl;
+            else
+                std::cout << "'" << key << "' :     " << coding[key] << std::endl;
+        }
+        std::cout << std::endl;
+
         int char_arr_size = (encoded_msg_len/8)+1;
         std::cout << "char array size: " << char_arr_size << std::endl;
-        for (int i=0; i < (encoded_msg_len/8)+1; i++) {
+        std::cout << "encoded message length: " << encoded_msg_len << " bits" << std::endl;
+        std::cout << "printing encoded message's array... " << std::endl;
+        for (int i=0; i < (int)(encoded_msg_len/8)+1; i++) {
             printbits(encoded_msg[i]);
+            std::cout << " ";
         }
     }
 
+    LastSavedSize = encoded_msg_len;
     return std::make_pair((unsigned int)encoded_msg_len, encoded_msg);
 }
 
-/* 0 = left, 1 = right
-*/
-std::string HuffmanCoding::bit_decode(unsigned char* encoded_msg, size_t n_bits, Node* tree_head) 
+std::string Huffman::bit_decode(unsigned char* encoded_msg, size_t n_bits, Node* tree_head) 
 {
     int EncMsg_idx = 0;
     std::string RetStr;
@@ -175,102 +157,6 @@ std::string HuffmanCoding::bit_decode(unsigned char* encoded_msg, size_t n_bits,
     }
 
     return RetStr;
-}
-
-/* -- Implementación funciones auxiliares -- */
-
-Node * combineNodes(Node * node1, Node * node2) {
-    Node * newnode = new Node{'~', node1->frequency + node2->frequency, node1, node2};
-    return newnode;
-}
-
-/*
-* Parámetros:
-* - textfin: fstream desde el cual se lee el archivo
-*
-* Retorna:
-* - unordered_map, cuyas claves son los carácteres del archivo, y cuyos valores son la frecuencia asociada al carácter dentro del texto.
-*/
-std::unordered_map<char, int> readFrequencies(std::fstream& textfin) {
-    std::unordered_map<char, int> counter;
-    char ch;
-    while (textfin >> std::noskipws >> ch) {
-        counter[ch]++;
-    }
-    return counter;
-}
-
-std::priority_queue<Node *, std::vector<Node*>, nodeComparator> createLeaves(const std::unordered_map<char, int>& map) {
-    std::priority_queue<Node *, std::vector<Node*>, nodeComparator> queue;
-    for (std::pair<char, int> pair: map) {
-        queue.push(new Node{pair.first, pair.second, nullptr, nullptr});
-    }
-    return queue;
-}
-
-Node * createHuffmanTree(std::unordered_map<char, int> freq) {
-    auto queue = createLeaves(freq);
-    while (queue.size() > 1) {
-        auto node1 = queue.top(); queue.pop();
-        auto node2 = queue.top(); queue.pop();
-        queue.push(combineNodes(node1, node2));
-    }
-    return queue.top(); /* Head */
-}
-
-bool searchValue(Node * node, char target, std::string& path) {
-    if (!node) {
-        return false;
-    }
-    if (node->symbol == target) {
-        return true;
-    }
-
-    /* Left */
-    path.push_back('0');
-    bool cond1 = searchValue(node->left, target, path);
-    if (cond1) {return true;}
-    path.pop_back();
-    
-    /* Right */
-    path.push_back('1');
-    bool cond2 = searchValue(node->right, target, path);
-    if (cond2) {return true;}
-    path.pop_back();
-
-    return false;
-}
-
-std::pair<std::unordered_map<char, std::string>, std::unordered_map<std::string, char>> getMaps(Node * huff, std::unordered_map<char, int> map) {
-    std::unordered_map<char, std::string> code;
-    std::unordered_map<std::string, char> decode;
-
-    for (auto it = map.begin(); it != map.end(); it++) {
-        auto key = it->first;
-        std::string path;
-        if (!searchValue(huff, key, path)) {
-            exit(EXIT_FAILURE); /* Siempre deberia encontrar un camino */
-        }
-        code.insert({key, path});
-        decode.insert({path, key});
-    }
-
-    return std::make_pair(code, decode);
-}
-
-/* -- Deprecated -- */
-
-void _inOrderTraversal(Node * node, int level) {
-    if (node != nullptr) {
-            _inOrderTraversal(node->left, level + 1);
-            std::cout << level << "\t" << node->symbol << "\t" << node->frequency << std::endl;
-            _inOrderTraversal(node->right, level + 1);
-        }
-}
-
-void inOrderTraversal(Node * head) {
-    std::cout << "Level\tSymbol\tFrequency" << std::endl;
-    _inOrderTraversal(head, 0);
 }
 
 /* debug */
